+++ app-side/login/LoginProvider.js
@@ -1,7 +1,4 @@
-import {GoogleAuth} from "./GoogleAuth";
-import {MicrosoftAuth} from "./MicrosoftAuth";
 import {CalDAVAuth} from "./CalDAVAuth";
-import {TickTickAuth} from "./TickTickAuth";
 
 export class LoginProvider {
     _getAuthProviderName() {
@@ -10,12 +7,6 @@ export class LoginProvider {
 
     _getAuthProvider() {
         switch(this._getAuthProviderName()) {
-            case "google":
-                return new GoogleAuth();
-            case "microsoft":
-                return new MicrosoftAuth();
-            case "tick_tick":
-                return new TickTickAuth();
             case "caldav":
                 return new CalDAVAuth();
         }